frontend-app-api: move createExtensionInstance to graph
