+++ packages/frontend-app-api/src/wiring/graph/createExtensionInstance.test.ts
