+++ packages/frontend-app-api/src/wiring/graph/createExtensionInstance.ts
